+++ CHANGELOG.md
@@ -9,6 +9,10 @@ See [`UPGRADE.md`](./UPGRADE.md) for additional help when upgrading to newer ver
 
 ## [Unreleased]
 
+### Added
+
+- `NifException` for using Elixir exception structs
+
 ### Fixed
 
 - Fix codegen problem for untagged enums (#370)